+++ lib/java/os-environment/src/test/java/org/enso/os/environment/jni/LoadClassTest.java
@@ -2,23 +2,24 @@ package org.enso.os.environment.jni;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.fail;
 
 import java.io.File;
-import java.util.HashMap;
-import java.util.Map;
+import java.math.BigInteger;
+import java.nio.file.Files;
 import java.util.Random;
 import org.enso.os.environment.jni.JNI.JValue;
-import org.graalvm.nativeimage.CurrentIsolate;
-import org.graalvm.nativeimage.IsolateThread;
 import org.graalvm.nativeimage.StackValue;
-import org.graalvm.nativeimage.c.function.CEntryPoint;
-import org.graalvm.nativeimage.c.function.CEntryPointLiteral;
-import org.graalvm.nativeimage.c.function.CFunctionPointer;
-import org.graalvm.nativeimage.c.type.CCharPointer;
 import org.graalvm.nativeimage.c.type.CTypeConversion;
+import org.junit.Before;
 import org.junit.Test;
 
 public class LoadClassTest {
+  // TBD: Make the number bigger again
+  //   - which currently exceeds the maximum size of a message
+  //   - so fix it
+  private static final int MAX = 5; // 5000;
+  private static final int MIN = 1; // 1000;
   private static final String PATH = System.getProperty("java.home");
   // set from TestCollectorFeature
   public static String MODULE_PATH;
@@ -43,6 +44,13 @@ public class LoadClassTest {
 
   private static JNI.JNIEnv env() {
     return jvm().env();
+  }
+
+  private Channel channel;
+
+  @Before
+  public void initializeChannel() throws Exception {
+    channel = Channel.create(jvm(), JVMPeer.class);
   }
 
   @Test
@@ -116,35 +124,108 @@ public class LoadClassTest {
 
   @Test
   public void executeMainClass() throws Exception {
-    var jvmIsolate = CurrentIsolate.getCurrentThread().rawValue();
-    var callbackFn = CALLBACK_FN.getFunctionPointer().rawValue();
+    var out = File.createTempFile("check-main", ".log");
+    var gen = new Random();
+    for (var i = 0; i < 5; i++) {
+      var n = gen.nextInt(MIN, MAX);
+      jvm().executeMain("org/enso/os/environment/jni/TestMain", out.getPath(), "" + n);
+      var content = Files.readString(out.toPath());
+      assertEquals("Factorial of " + n + " is the same", TestMain.factorial(n).toString(), content);
+      out.delete();
+    }
+  }
+
+  @Test
+  public void computeFactorialViaMessages() throws Exception {
+    TestMain.CORRECT_RESULTS.clear();
+    assertEquals("Results are empty", 0, TestMain.CORRECT_RESULTS.size());
     var gen = new Random();
     var n = 0L;
     for (var i = 0; i < 5; i++) {
-      n += gen.nextLong(1000, 5000);
-      var mainClass = "org/enso/os/environment/jni/TestMain";
-      jvm().executeMain(mainClass, "" + jvmIsolate, "" + callbackFn, "" + n);
+      n += gen.nextLong(MIN, MAX);
+      channel.execute(Void.class, new TestMain.RequestFactorial(n));
     }
-    assertEquals("Five results found: " + CORRECT_RESULTS, 5, CORRECT_RESULTS.size());
+    assertEquals(
+        "Five results found: " + TestMain.CORRECT_RESULTS, 5, TestMain.CORRECT_RESULTS.size());
+    for (var e : TestMain.CORRECT_RESULTS.entrySet()) {
+      var expecting = TestMain.factorial(e.getKey());
+      assertEquals("fac(" + e.getKey() + ") should be", expecting.toString(), e.getValue());
+    }
   }
 
-  private static final Map<Long, String> CORRECT_RESULTS = new HashMap<>();
-
-  @CEntryPoint
-  private static boolean acceptResultFromHotSpotJvm(
-      IsolateThread threadId, long n, CCharPointer resultStr) {
-    var result = CTypeConversion.toJavaString(resultStr);
-    var ownResult = TestMain.factorial(n).toString();
-    assertEquals("fac(" + n + ") is correct in both JVMs", ownResult, result);
-    CORRECT_RESULTS.put(n, result);
-    return ownResult.equals(result);
+  @Test
+  public void computeFactorialViaSingleMessage() throws Exception {
+    var gen = new Random();
+    var n = 0L;
+    for (var i = 0; i < 5; i++) {
+      n += gen.nextLong(MIN, MAX);
+      var res = channel.execute(BigInteger.class, new TestMain.ComputeFactorial(n));
+      var expecting = TestMain.factorial(n);
+      assertEquals("fac(" + n + ") should be", expecting, res);
+    }
   }
 
-  private static final CEntryPointLiteral<CFunctionPointer> CALLBACK_FN =
-      CEntryPointLiteral.create(
-          LoadClassTest.class,
-          "acceptResultFromHotSpotJvm",
-          IsolateThread.class,
-          long.class,
-          CCharPointer.class);
+  @Test
+  public void backAndForthFactorialOne() throws Exception {
+    var fac = channel.execute(Long.class, new TestMain.CountDownAndReturn(1, 1));
+    assertEquals(1, fac.longValue());
+  }
+
+  @Test
+  public void backAndForthFactorialTwo() throws Exception {
+    var fac = channel.execute(Long.class, new TestMain.CountDownAndReturn(2, 1));
+    assertEquals(2, fac.longValue());
+  }
+
+  @Test
+  public void backAndForthFactorialThree() throws Exception {
+    var fac = channel.execute(Long.class, new TestMain.CountDownAndReturn(3, 1));
+    assertEquals(6, fac.longValue());
+  }
+
+  @Test
+  public void backAndForthFactorialFour() throws Exception {
+    var fac = channel.execute(Long.class, new TestMain.CountDownAndReturn(4, 1));
+    assertEquals(24, fac.longValue());
+  }
+
+  @Test
+  public void backAndForthFactorialFive() throws Exception {
+    var fac = channel.execute(Long.class, new TestMain.CountDownAndReturn(5, 1));
+    assertEquals(120, fac.longValue());
+  }
+
+  @Test
+  public void throwFactorialOne() throws Exception {
+    assertException("1", new TestMain.CountDownAndThrow(1, 1));
+  }
+
+  @Test
+  public void throwFactorialTwo() throws Exception {
+    assertException("2", new TestMain.CountDownAndThrow(2, 1));
+  }
+
+  @Test
+  public void throwFactorialThree() throws Exception {
+    assertException("6", new TestMain.CountDownAndThrow(3, 1));
+  }
+
+  @Test
+  public void throwFactorialFour() throws Exception {
+    assertException("24", new TestMain.CountDownAndThrow(4, 1));
+  }
+
+  @Test
+  public void throwFactorialFive() throws Exception {
+    assertException("120", new TestMain.CountDownAndThrow(5, 1));
+  }
+
+  private void assertException(String msg, TestMain.CountDownAndThrow action) {
+    try {
+      channel.execute(Void.class, action);
+      fail("Expecting an exception to be thrown for " + msg);
+    } catch (IllegalStateException ex) {
+      assertEquals(msg, ex.getMessage());
+    }
+  }
 }