+++ std-bits/table/src/main/java/org/enso/table/data/column/builder/DoubleBuilder.java
@@ -2,6 +2,7 @@ package org.enso.table.data.column.builder;
 
 import java.math.BigDecimal;
 import java.math.BigInteger;
+import java.util.BitSet;
 import java.util.Objects;
 import org.enso.base.polyglot.NumericConverter;
 import org.enso.table.data.column.storage.BoolStorage;
@@ -22,6 +23,13 @@ import org.enso.table.util.BitSets;
 
 /** A builder for floating point columns. */
 public class DoubleBuilder extends NumericBuilder implements BuilderForDouble {
+  public static Storage<Double> makeEmpty(long size) {
+    int intSize = Builder.checkSize(size);
+    BitSet isNothing = new BitSet(intSize);
+    isNothing.set(0, intSize);
+    return new DoubleStorage(new double[0], intSize, isNothing);
+  }
+
   protected final PrecisionLossAggregator precisionLossAggregator;
   protected double[] data;