+++ std-bits/table/src/main/java/org/enso/table/data/column/operation/comparators/NumericComparators.java
@@ -2,11 +2,12 @@ package org.enso.table.data.column.operation.comparators;
 
 import java.math.BigDecimal;
 import java.math.BigInteger;
+import org.enso.table.data.column.builder.BoolBuilder;
 import org.enso.table.data.column.operation.BinaryOperation;
 import org.enso.table.data.column.operation.BinaryOperationNumeric;
+import org.enso.table.data.column.operation.NumericColumnAdapter;
 import org.enso.table.data.column.operation.StorageIterators;
 import org.enso.table.data.column.operation.map.MapOperationProblemAggregator;
-import org.enso.table.data.column.storage.BoolStorage;
 import org.enso.table.data.column.storage.ColumnDoubleStorage;
 import org.enso.table.data.column.storage.ColumnLongStorage;
 import org.enso.table.data.column.storage.ColumnStorage;
@@ -227,16 +228,16 @@ abstract class NumericComparators<T> extends BinaryOperationNumeric<T, Boolean> 
   @Override
   protected ColumnStorage<Boolean> applyNullMap(
       ColumnStorage<?> left, MapOperationProblemAggregator problemAggregator) {
-    return BoolStorage.makeEmpty(left.getSize());
+    return BoolBuilder.makeEmpty(left.getSize());
   }
 
   private static class NumericComparatorsDouble extends NumericComparators<Double> {
     public NumericComparatorsDouble(NumericComparator comparator) {
-      super(DoubleColumnAdapter.INSTANCE, comparator);
+      super(NumericColumnAdapter.DoubleColumnAdapter.INSTANCE, comparator);
     }
 
     public NumericComparatorsDouble(NumericComparator comparator, boolean valueOnOther) {
-      super(DoubleColumnAdapter.INSTANCE, comparator, valueOnOther);
+      super(NumericColumnAdapter.DoubleColumnAdapter.INSTANCE, comparator, valueOnOther);
     }
 
     @Override
@@ -275,11 +276,11 @@ abstract class NumericComparators<T> extends BinaryOperationNumeric<T, Boolean> 
 
   private static class NumericComparatorsBigDecimal extends NumericComparators<BigDecimal> {
     public NumericComparatorsBigDecimal(NumericComparator comparator) {
-      super(BigDecimalColumnAdapter.INSTANCE, comparator);
+      super(NumericColumnAdapter.BigDecimalColumnAdapter.INSTANCE, comparator);
     }
 
     public NumericComparatorsBigDecimal(NumericComparator comparator, boolean valueOnOther) {
-      super(BigDecimalColumnAdapter.INSTANCE, comparator, valueOnOther);
+      super(NumericColumnAdapter.BigDecimalColumnAdapter.INSTANCE, comparator, valueOnOther);
     }
 
     @Override
@@ -294,11 +295,11 @@ abstract class NumericComparators<T> extends BinaryOperationNumeric<T, Boolean> 
 
   private static class NumericComparatorsBigInteger extends NumericComparators<BigInteger> {
     public NumericComparatorsBigInteger(NumericComparator comparator) {
-      super(BigIntegerColumnAdapter.INSTANCE, comparator);
+      super(NumericColumnAdapter.BigIntegerColumnAdapter.INSTANCE, comparator);
     }
 
     public NumericComparatorsBigInteger(NumericComparator comparator, boolean valueOnOther) {
-      super(BigIntegerColumnAdapter.INSTANCE, comparator, valueOnOther);
+      super(NumericColumnAdapter.BigIntegerColumnAdapter.INSTANCE, comparator, valueOnOther);
     }
 
     @Override
@@ -313,11 +314,11 @@ abstract class NumericComparators<T> extends BinaryOperationNumeric<T, Boolean> 
 
   private static class NumericComparatorsLong extends NumericComparators<Long> {
     public NumericComparatorsLong(NumericComparator operation) {
-      super(LongColumnAdapter.INSTANCE, operation);
+      super(NumericColumnAdapter.LongColumnAdapter.INSTANCE, operation);
     }
 
     public NumericComparatorsLong(NumericComparator operation, boolean valueOnOther) {
-      super(LongColumnAdapter.INSTANCE, operation, valueOnOther);
+      super(NumericColumnAdapter.LongColumnAdapter.INSTANCE, operation, valueOnOther);
     }
 
     @Override