+++ std-bits/table/src/main/java/org/enso/table/data/column/storage/numeric/AbstractLongStorage.java
@@ -5,19 +5,12 @@ import java.util.List;
 import org.enso.table.data.column.builder.Builder;
 import org.enso.table.data.column.operation.map.MapOperationProblemAggregator;
 import org.enso.table.data.column.operation.map.MapOperationStorage;
-import org.enso.table.data.column.operation.map.numeric.LongRoundOp;
 import org.enso.table.data.column.operation.map.numeric.arithmetic.AddOp;
 import org.enso.table.data.column.operation.map.numeric.arithmetic.DivideOp;
 import org.enso.table.data.column.operation.map.numeric.arithmetic.ModOp;
 import org.enso.table.data.column.operation.map.numeric.arithmetic.MulOp;
 import org.enso.table.data.column.operation.map.numeric.arithmetic.PowerOp;
 import org.enso.table.data.column.operation.map.numeric.arithmetic.SubOp;
-import org.enso.table.data.column.operation.map.numeric.comparisons.EqualsComparison;
-import org.enso.table.data.column.operation.map.numeric.comparisons.GreaterComparison;
-import org.enso.table.data.column.operation.map.numeric.comparisons.GreaterOrEqualComparison;
-import org.enso.table.data.column.operation.map.numeric.comparisons.LessComparison;
-import org.enso.table.data.column.operation.map.numeric.comparisons.LessOrEqualComparison;
-import org.enso.table.data.column.operation.map.numeric.isin.LongIsInOp;
 import org.enso.table.data.column.storage.*;
 import org.enso.table.data.column.storage.type.IntegerType;
 import org.enso.table.data.column.storage.type.StorageType;
@@ -27,7 +20,18 @@ import org.enso.table.problems.BlackholeProblemAggregator;
 import org.graalvm.polyglot.Context;
 
 public abstract class AbstractLongStorage extends Storage<Long> implements ColumnLongStorage {
-  private static final MapOperationStorage<Long, AbstractLongStorage> ops = buildOps();
+  private static final MapOperationStorage<Long, AbstractLongStorage> OPS = buildOps();
+
+  private static MapOperationStorage<Long, AbstractLongStorage> buildOps() {
+    MapOperationStorage<Long, AbstractLongStorage> ops = new MapOperationStorage<>();
+    ops.add(new AddOp<>())
+        .add(new SubOp<>())
+        .add(new MulOp<>())
+        .add(new DivideOp<>())
+        .add(new ModOp<>())
+        .add(new PowerOp<>());
+    return ops;
+  }
 
   private final long size;
   private final IntegerType type;
@@ -59,34 +63,15 @@ public abstract class AbstractLongStorage extends Storage<Long> implements Colum
   public abstract long getItemAsLong(long index) throws ValueIsNothingException;
 
   @Override
-  public boolean isBinaryOpVectorized(String name) {
-    return ops.isSupportedBinary(name);
-  }
-
-  @Override
-  public Storage<?> runVectorizedBinaryMap(
+  protected Storage<?> runVectorizedBinaryMap(
       String name, Object argument, MapOperationProblemAggregator problemAggregator) {
-    return ops.runBinaryMap(name, this, argument, problemAggregator);
+    return OPS.runBinaryMap(name, this, argument, problemAggregator);
   }
 
   @Override
-  public boolean isTernaryOpVectorized(String op) {
-    return ops.isSupportedTernary(op);
-  }
-
-  @Override
-  public Storage<?> runVectorizedTernaryMap(
-      String name,
-      Object argument0,
-      Object argument1,
-      MapOperationProblemAggregator problemAggregator) {
-    return ops.runTernaryMap(name, this, argument0, argument1, problemAggregator);
-  }
-
-  @Override
-  public Storage<?> runVectorizedZip(
+  protected Storage<?> runVectorizedZip(
       String name, Storage<?> argument, MapOperationProblemAggregator problemAggregator) {
-    return ops.runZip(name, this, argument, problemAggregator);
+    return OPS.runZip(name, this, argument, problemAggregator);
   }
 
   @Override
@@ -140,24 +125,6 @@ public abstract class AbstractLongStorage extends Storage<Long> implements Colum
     }
 
     return possibleTypes[currentTypeIdx];
-  }
-
-  private static MapOperationStorage<Long, AbstractLongStorage> buildOps() {
-    MapOperationStorage<Long, AbstractLongStorage> ops = new MapOperationStorage<>();
-    ops.add(new AddOp<>())
-        .add(new SubOp<>())
-        .add(new MulOp<>())
-        .add(new DivideOp<>())
-        .add(new ModOp<>())
-        .add(new PowerOp<>())
-        .add(new LongRoundOp(Maps.ROUND))
-        .add(new LessComparison<>())
-        .add(new LessOrEqualComparison<>())
-        .add(new EqualsComparison<>())
-        .add(new GreaterOrEqualComparison<>())
-        .add(new GreaterComparison<>())
-        .add(new LongIsInOp());
-    return ops;
   }
 
   @Override